Move normal outputting to text formatter
